requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.16,<3.17
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.28.0,<0.29.0
django-cors-headers>=4.7.0
aiohappyeyeballs>=2.6.1
aiohttp>=3.11.18
aiosignal>=1.3.2
attrs>=25.3.0
beautifulsoup4>=4.13.3
certifi>=2025.1.31
charset-normalizer>=3.4.1
click>=8.1.8
colorama>=0.4.6
frozenlist>=1.6.0
idna>=3.10
joblib>=1.4.2
lxml>=5.4.0
multidict>=6.4.3
nltk>=3.9.1
numpy>=2.0.2
propcache>=0.3.1
regex>=2024.11.6
requests>=2.32.3
scikit-learn>=1.6.1
scipy>=1.13.1
soupsieve>=2.6
threadpoolctl>=3.5.0
tqdm>=4.67.1
typing_extensions>=4.12.2
urllib3>=2.3.0
yarl>=1.20.0
apify-client>=1.10.0